// src/enums/entityTypes.ts
export const enum EntityPlural {
    TRACKS = 'tracks',
    BANDS = 'bands',
    ARTISTS = 'artists',
    GENRES = 'genres',
    ALBUMS = 'albums',
    FAVOURITES = 'favourites',
    USERS = 'users'
}

export const enum Entity {
    TRACK = 'track',
    BAND = 'band',
    ARTIST = 'artist',
    GENRE = 'genre',
    ALBUM = 'album',
    FAVOURITE = 'favourite',
    USER = 'user'
}